% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{calc_fit_errors}
\alias{calc_fit_errors}
\title{Calculate the fitting error}
\usage{
calc_fit_errors(data.type, data, dfsim, nag)
}
\value{
Dataframe of errors, corresponding prior indices and variable name fitted.
}
\description{
Calculate the fitting error
}
\keyword{internal}
