% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_data_fit}
\alias{check_data_fit}
\title{Check the integrity of the data to fit.}
\usage{
check_data_fit(data)
}
\arguments{
\item{data}{List of dataframe.}
}
\value{
An error if the data integrity checks fail.
}
\description{
Check the integrity of the data to fit.
}
