% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_data_fit_date}
\alias{check_data_fit_date}
\title{Check dates consistency between data to fit and model parameters}
\usage{
check_data_fit_date(data, obj)
}
\arguments{
\item{data}{List of data as expected by function \code{amrem::fit()}.}

\item{obj}{List defining an \code{amrem} object.}
}
\value{
An error if relevant.
}
\description{
Check dates consistency between data to fit and model parameters
}
\keyword{internal}
