% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_first_letter}
\alias{check_first_letter}
\title{Check the first element of character vector are letters.}
\usage{
check_first_letter(x)
}
\description{
Check the first element of character vector are letters.
}
\keyword{internal}
