% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_name}
\alias{check_name}
\title{Check if a parameter name is present.}
\usage{
check_name(x, prms)
}
\arguments{
\item{x}{}

\item{prms}{}
}
\description{
Check if a parameter name is present.
}
\keyword{internal}
