% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_create}
\alias{check_prms_create}
\title{Check model parameters before object creation}
\usage{
check_prms_create(prms)
}
\arguments{
\item{prms}{List of model parameters}
}
\value{
Nothing
}
\description{
Check model parameters before object creation
}
