% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_create_i0}
\alias{check_prms_create_i0}
\title{Check initial infected individuals parameters}
\usage{
check_prms_create_i0(prms)
}
\arguments{
\item{prms}{List of model parameters}
}
\value{
Nothing
}
\description{
Check initial infected individuals parameters
}
\keyword{internal}
