% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_dist_matrix}
\alias{check_prms_dist_matrix}
\title{Check if the parameters of `dist_create_matrix()` are valid.}
\usage{
check_prms_dist_matrix(means, vars, maxs)
}
\arguments{
\item{means}{matrix of means}

\item{vars}{matrix of variances}

\item{maxs}{matrix of maximums}
}
\description{
Check if the parameters of `dist_create_matrix()` are valid.
}
\keyword{internal}
