% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_fit}
\alias{check_prms_fit}
\title{Check the argument parameters for fit().}
\usage{
check_prms_fit(prms.fit, nag)
}
\arguments{
\item{prms.fit}{List of parameters}

\item{nag}{Integer. Number of age groups.}
}
\description{
Check the argument parameters for fit().
}
\keyword{internal}
