% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_i0}
\alias{check_prms_i0}
\title{Check i0 specifically as it can be
specified as either a proportion `i0.prop`
or as an incidence matrix `i0`.}
\usage{
check_prms_i0(prms)
}
\arguments{
\item{prms}{List of model parameters}
}
\value{
List of (potentially modified) model parameters.
}
\description{
Check i0 specifically as it can be
specified as either a proportion `i0.prop`
or as an incidence matrix `i0`.
}
