% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_name}
\alias{check_prms_name}
\title{Check if a parameter name is present.}
\usage{
check_prms_name(x, prms)
}
\arguments{
\item{x}{Name of parameter}

\item{prms}{List of parameters.}
}
\description{
Check if a parameter name is present.
}
\keyword{internal}
