% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_S0}
\alias{check_prms_S0}
\title{Ensure parameter `S0` is consistent
when `S0.prop` is specified (instead of `S0` directly).}
\usage{
check_prms_S0(prms)
}
\arguments{
\item{prms}{List of model parameters}
}
\value{
List of (potentially modified) model parameters.
}
\description{
Ensure parameter `S0` is consistent
when `S0.prop` is specified (instead of `S0` directly).
}
