% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_prms_simulate}
\alias{check_prms_simulate}
\title{Check the parameters for `simulate()` is valid.}
\usage{
check_prms_simulate(prms)
}
\arguments{
\item{prms}{List of model parameters.}
}
\value{
NULL if valid, an error if invalid.
}
\description{
Check the parameters for `simulate()` is valid.
}
\keyword{internal}
