% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{digest_long_data}
\alias{digest_long_data}
\title{Transform a long-format dataframe  
into a list of dataframes.}
\usage{
digest_long_data(data, dictionary)
}
\arguments{
\item{data}{Dataframe in a long format.}

\item{dictionary}{Dataframe providing the link
between the variable names of the input data
and the standard ones expected by \code{amrem}.}
}
\value{
List of data frames. One element for each 
variable and age group (e.g., \code{testpos_1}, 
\code{hospadm_2}, etc.)
}
\description{
Transform a long-format dataframe  
into a list of dataframes.
}
\examples{
 data = example_longformat_data()
 dictionary = data.frame(
amrem = c('testpos', 'hosp'),
name = c('positivity', 'hosp') )



}
