% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.R
\name{dist_create}
\alias{dist_create}
\title{Create a Gamma distribution.}
\usage{
dist_create(mean, var, max)
}
\arguments{
\item{mean}{Numeric. Mean of the distribution.}

\item{var}{Numeric. Variance of the distribution.}

\item{max}{Integer. Maximum support of the distribution.}
}
\value{
Numerical vector. Normalized (sum = 1).
}
\description{
Create a Gamma distribution.
}
\examples{
dist_create(mean=3, var=1.2, max=5)


}
