% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.R
\name{dist_create_matrix}
\alias{dist_create_matrix}
\title{Create a matrix of Gamma distributions.}
\usage{
dist_create_matrix(means, vars, maxs)
}
\arguments{
\item{means}{Numeric matrix. Means of the distribution.}

\item{vars}{Numeric matrix. Variances of the distribution.}

\item{maxs}{Integer matrix. Maximums support of the distribution.}
}
\value{
List of list of distribution. Element \code{[[i]][[j]]} is the distribution with mean \code{means[i,j]}.
}
\description{
Create a matrix of Gamma distributions.
}
\examples{
 means = matrix(c(2,3,2,2), ncol = 2)
 vars  = matrix(c(1.1, 1.8, 1.4, 0.9), ncol = 2)
 maxs  = matrix(c(5,5,5,6), ncol = 2)
 dist_create_matrix(means, vars, maxs)

}
