% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_longformat_data}
\alias{example_longformat_data}
\title{Data example, with long format.}
\usage{
example_longformat_data(n = 12)
}
\arguments{
\item{n}{Integer. Number of data points by varable and age group.}
}
\value{
List of data frame as expected by \code{amrem}.
}
\description{
Data example, with long format.
}
\examples{
data = example_longformat_data()


}
