% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_model_prms}
\alias{example_model_prms}
\title{Example of model parameters}
\usage{
example_model_prms()
}
\value{
List Example of model parameters
}
\description{
Example of model parameters
}
\examples{
prms = example_model_prms()

}
