% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_model_prms_ag}
\alias{example_model_prms_ag}
\title{Example of model parameters with flexible number of age groups.}
\usage{
example_model_prms_ag(
  r0 = 1.3,
  N = c(0.7, 2.4, 4.7, 3.9, 2.7) * 1e+06,
  S0.prop = c(0.99, 0.8, 0.3, 0.3, 0.6),
  h.prop = c(0.02, 0.01, 0.001, 0.03, 0.05),
  odds.testpos = c(3, 2, 1, 5, 10),
  R_decay = 0.1,
  R_min = 0.1,
  R_max = 0.5,
  fec.scale = 1
)
}
\arguments{
\item{r0}{Numeric. Global multiplier for contact matrix.}

\item{N}{Numeric vector of population sizes.}

\item{S0.prop}{Numeric vector of proportion of initial susceptible populations.}

\item{h.prop}{Numeric vector of proportion of hospitalization rate for each age group.}

\item{odds.testpos}{Numeric vector of odds to test positive for each age group.}
}
\value{
List Example of model parameters
}
\description{
Example of model parameters with flexible number of age groups.
}
\examples{
prms = example_model_prms_ag()

}
