% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_simulated_data}
\alias{example_simulated_data}
\title{Create synthetic data from a simulated epidemic}
\usage{
example_simulated_data(model.prms, date.obs)
}
\arguments{
\item{model.prms}{List. Model parameters as returned by the function \code{example_model_prms()}.}

\item{date.obs}{Vector of observation dates.}
}
\value{
List of data sets as expected by the function \code{amrem::fit()}.
}
\description{
Create synthetic data from a simulated epidemic
}
\examples{

model.prms = example_model_prms()
t.obs = 12*c(1:10)    # observation times
date.obs = model.prms$date.start + 12*c(1:10)    # observation dates
dat = example_simulated_data(model.prms, date.obs)

}
