% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a model object to data.}
\usage{
fit(obj, prms.fit, data)
}
\arguments{
\item{obj}{List Model object as returned by \code{amrem::create()}.}

\item{prms.fit}{List Parameters related to the fitting algorithm.}

\item{data}{List Data to fit the model to.}
}
\value{
TODO
}
\description{
Fit a model object to data.
}
\examples{
foo = fit()


}
