% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_data}
\alias{flatten_data}
\title{Flatten data structure}
\usage{
flatten_data(data)
}
\arguments{
\item{data}{List of dataframe having the same structure as 
the list of data passed as argument of the function \code{fit()}.}
}
\value{
A single, merged and long-format dataframe.
}
\description{
Flatten data structure
}
