% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{generate_priors}
\alias{generate_priors}
\title{Generate priors}
\usage{
generate_priors(priors.dist, nag)
}
\arguments{
\item{priors.dist}{List defining the priors.}

\item{nag}{Integer. Number of age groups.}
}
\value{
List of priors arrays.
}
\description{
Generate priors
}
\keyword{internal}
