% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_colors_age_groups}
\alias{get_colors_age_groups}
\title{Return the color palette for age groups}
\usage{
get_colors_age_groups(nag)
}
\arguments{
\item{nag}{Integer. Number of age groups.}
}
\value{
Vector of colors.
}
\description{
Return the color palette for age groups
}
\examples{
get_colors_age_groups(3)

}
\keyword{internal}
