% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_names}
\alias{get_data_names}
\title{Helper function.
Generate the expected names of elements 
in the `data` list, including age group.}
\usage{
get_data_names(data.type, nag)
}
\arguments{
\item{data.type}{String.}

\item{nag}{Integer. Number of age groups.}
}
\description{
Helper function.
Generate the expected names of elements 
in the `data` list, including age group.
}
\keyword{internal}
