% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_last_date_data}
\alias{get_last_date_data}
\title{Helper function. Return the last date of each data sets.}
\usage{
get_last_date_data(fitobj)
}
\arguments{
\item{fitobj}{List object as returned by the function \code{fit()}.}
}
\value{
List of dates.
}
\description{
Helper function. Return the last date of each data sets.
}
\keyword{internal}
