% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nag}
\alias{get_nag}
\title{Return the number of age groups of a model}
\usage{
get_nag(obj)
}
\arguments{
\item{obj}{List. Object as returned by the function \code{amrem::create()}.}
}
\value{
Integer. Number of age groups.
}
\description{
Return the number of age groups of a model
}
\examples{

model.prms = amrem::example_model_prms()
obj = amrem::create(model.prms)
print(get_nag(obj))

}
