% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_matrix_type}
\alias{get_param_matrix_type}
\title{Returns the parameters considered as matrices}
\usage{
get_param_matrix_type()
}
\value{
Character vector of parameter names
}
\description{
Returns the parameters considered as matrices
}
