% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_vector_type}
\alias{get_param_vector_type}
\title{Returns the parameters considered as vectors}
\usage{
get_param_vector_type()
}
\value{
Character vector of parameter names
}
\description{
Returns the parameters considered as vectors
}
