% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sim_varnames}
\alias{get_sim_varnames}
\title{Helper function.
Link data names with simulation variables names.}
\usage{
get_sim_varnames(data.type, nag)
}
\arguments{
\item{data.type}{String.}

\item{nag}{Integer for the number of age groups.}
}
\description{
Helper function.
Link data names with simulation variables names.
}
\examples{
get_sim_varnames('hospadm', 2) 
# returns: c("h_1", "h_2")

}
\keyword{internal}
