% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{helper_fit_data_type}
\alias{helper_fit_data_type}
\title{Helper function for type of data fit}
\usage{
helper_fit_data_type(v, nam, x)
}
\arguments{
\item{v}{String. Name of variable.}

\item{nam}{String vector. Names of all variables.}

\item{x}{List.}
}
\description{
Helper function for type of data fit
}
\keyword{internal}
