% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helper_generate_gi}
\alias{helper_generate_gi}
\title{Helper to generate GI distributions}
\usage{
helper_generate_gi(A, gi.mean = 3, gi.var = 1, gi.max = 7)
}
\description{
Helper to generate GI distributions
}
\keyword{internal}
