% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{helper_list_to_array}
\alias{helper_list_to_array}
\title{Helper to convert a list into an array}
\usage{
helper_list_to_array(x)
}
\arguments{
\item{x}{List}
}
\value{
Array
}
\description{
Helper to convert a list into an array
}
\keyword{internal}
