% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{helper_listmat_to_df}
\alias{helper_listmat_to_df}
\title{Helper function to Convert a
list of matrices into a dataframe.}
\usage{
helper_listmat_to_df(x, pname)
}
\arguments{
\item{x}{List of matrices.}

\item{pname}{String. Parameter name.}
}
\description{
Helper function to Convert a
list of matrices into a dataframe.
}
