% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{helper_priors_mat}
\alias{helper_priors_mat}
\title{Helper function for prior matrices.}
\usage{
helper_priors_mat(varname, nag, priors.dist)
}
\arguments{
\item{varname}{String. Variable name.}

\item{nag}{Integer. Number of age groups.}

\item{priors.dist}{List defining prior distributions.}
}
\description{
Helper function for prior matrices.
}
\keyword{internal}
