% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{helper_priors_mat2}
\alias{helper_priors_mat2}
\title{Helper function to generate priors
for a parameter defined a matrix format}
\usage{
helper_priors_mat2(varname, nag, priors.dist, n.priors)
}
\arguments{
\item{varname}{String. Name of the variable.}

\item{nag}{Integer. Number of age groups.}

\item{priors.dist}{List defining the prior 
distribution for each matrix element.}

\item{n.priors}{Integer. Number of prior samples.}
}
\value{
Array of prior values.
}
\description{
Helper function to generate priors
for a parameter defined a matrix format
}
\keyword{internal}
