% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{helper_summarise_post}
\alias{helper_summarise_post}
\title{Helper function to summarise posterior data.}
\usage{
helper_summarise_post(df, ci)
}
\arguments{
\item{df}{dataframe of posteriors.}

\item{ci}{width of credible interval}
}
\value{
dataframe of summary statistics
}
\description{
Helper function to summarise posterior data.
}
\keyword{internal}
