% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{normalize_minmax}
\alias{normalize_minmax}
\title{Normalize element values of a vector to [0,1] using its min and max values.}
\usage{
normalize_minmax(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Normalized vector.
}
\description{
Normalize element values of a vector to [0,1] using its min and max values.
}
