% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data_list}
\alias{plot_data_list}
\title{Plot data inputted in list format.}
\usage{
plot_data_list(data)
}
\arguments{
\item{data}{List of data (as expected by the function \code{amrem::fit()}.)}
}
\value{
A ggplot object
}
\description{
Plot data inputted in list format.
}
\examples{

model.prms = example_model_prms()
t.obs = 12*c(1:10)    # observation times
date.obs = model.prms$date.start + 12*c(1:10)    # observation dates
data = example_simulated_data(model.prms, date.obs)  
g = plot_data_list(data)
plot(g)

}
