% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data_long}
\alias{plot_data_long}
\title{Plot data entered as a long dataframe}
\usage{
plot_data_long(data)
}
\arguments{
\item{data}{Dataframe in long format. Must have `source`, `date`, and `value` columns.
The column `source` must identify the age group by appending it. 
For example, `testpos_1`, `testpos_2`, `hospadm_1`, etc.}
}
\value{
A ggplot.
}
\description{
Plot data entered as a long dataframe
}
