% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_dist_matrix}
\alias{plot_dist_matrix}
\title{Plot a matrix of distributions}
\usage{
plot_dist_matrix(d)
}
\arguments{
\item{d}{List of distribution}
}
\value{
plot
}
\description{
Plot a matrix of distributions
}
