% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit_errors}
\alias{plot_fit_errors}
\title{Plot errors}
\usage{
plot_fit_errors(fitobj)
}
\arguments{
\item{fitobj}{Object as returned by \code{fit()}.}
}
\value{
list of ggplot objects
}
\description{
Plot errors
}
