% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit_post_2d}
\alias{plot_fit_post_2d}
\title{Plot 2D posterior distributions}
\usage{
plot_fit_post_2d(fitobj, true.values = NULL)
}
\arguments{
\item{fitobj}{List. Fit object as returned by the function \code{\link{fit}()}.}

\item{true.values}{Not used.}
}
\value{
List of ggplot objects.
}
\description{
Plot 2D posterior distributions
}
