% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit_post_vec}
\alias{plot_fit_post_vec}
\title{Plot posterior distributions for vector parameters.}
\usage{
plot_fit_post_vec(prmname, prior, post, ci, true.values)
}
\arguments{
\item{prmname}{String. Name of the parameter.}

\item{post}{Dataframe. Posterior data as returned by \code{fit(...)$post}.}

\item{ci}{Numeric. Width of the credible interval.}
}
\value{
A ggplot object.
}
\description{
Plot posterior distributions for vector parameters.
}
\keyword{internal}
