% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit_traj}
\alias{plot_fit_traj}
\title{Plot fitted trajectories}
\usage{
plot_fit_traj(fitobj, ci = 0.95)
}
\arguments{
\item{fitobj}{List as returned by the functino \code{fit()}.}

\item{ci}{Numeric. Width of the quantiles to plot for the fitted trajectories.}
}
\value{
A ggplot object.
}
\description{
Plot fitted trajectories
}
