% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot simulated time series}
\usage{
plot_timeseries(sim)
}
\arguments{
\item{sim}{Dataframe. Simulation returned by function \code{amrem::simulate()}.}
}
\value{
A `ggplot` object.
}
\description{
Plot simulated time series
}
\examples{

prms = amrem::example_model_prms()
obj  = amrem::create(prms) 
sim  = amrem::simulate(obj)
g    = amrem::plot_timeseries(sim)
plot(g)

}
