% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{process_prms_i0prop}
\alias{process_prms_i0prop}
\title{Process `i0.prop` parameter}
\usage{
process_prms_i0prop(prms)
}
\arguments{
\item{prms}{List of model parameters.}
}
\value{
List of model parameters with `i0` matrix if `i0.prop` is present.
}
\description{
Process `i0.prop` parameter
}
\keyword{internal}
