% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate an epidemic.}
\usage{
simulate(obj, check.prms = TRUE)
}
\arguments{
\item{obj}{List representing the `amrem` object as returned by \code{amrem::create()}.}

\item{check.prms}{Logical. Check input parameters? Default = TRUE.}
}
\value{
Dataframe of simulated variables.
}
\description{
Simulate an epidemic.
}
\examples{
1 + 2 #TODO


}
