% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_c}
\alias{simulate_c}
\title{Simulate an epidemic model}
\usage{
simulate_c(prms, ww)
}
\arguments{
\item{prms}{A list of parameters including N, S0, horizon, alpha, g, i0, R.}

\item{ww}{A boolean to switch on/off fecal shedding into wastewater.}
}
\value{
A data frame with incidence and susceptible counts over time.
}
\description{
Runs a simulation using specified parameters.
}
\examples{
simulate_c(list(N = c(1000, 1000), S0 = c(990, 990), horizon = 50,
                   alpha = 0.1, g = c(0.5, 0.5), i0 = matrix(1, nrow = 2, ncol = 2),
                   R = matrix(c(1, 0.5, 0.5, 1), nrow = 2)))
}
