% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{simulate_fit_unit}
\alias{simulate_fit_unit}
\title{Simulation used in fit algorithm 
to calculate error from data}
\usage{
simulate_fit_unit(i, obj, priors, data)
}
\arguments{
\item{i}{Integer. Index}

\item{obj}{List. Model object.}

\item{priors}{Dataframe of priors.}
}
\value{
Dataframe of prior indices and errors.
}
\description{
Simulation used in fit algorithm 
to calculate error from data
}
\keyword{internal}
