% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{simulate_r}
\alias{simulate_r}
\title{Simulate epidemic.}
\usage{
simulate_r(prms)
}
\arguments{
\item{prms}{List of model parameters.}
}
\value{
Dataframe of simulated epidemic.
}
\description{
Simulate epidemic.
}
\examples{
simulate(list(foo = 1))

}
