% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_age_groups}
\alias{standardize_age_groups}
\title{Standardize labels for age groups}
\usage{
standardize_age_groups(x)
}
\arguments{
\item{x}{Character vector of age group labels}
}
\value{
Character vector of standardized age group labels.
}
\description{
Standardize labels for age groups
}
\examples{

 x = c('0-4', '0 to 4', '65+', '65 +')
 standardize_age_groups(x)


}
