% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{summarize_forecast}
\alias{summarize_forecast}
\title{Summary statistics if a forecast object.}
\usage{
summarize_forecast(fcstobj, ci)
}
\arguments{
\item{fcstobj}{List. Object as returned by the function \code{forecast(...)}.}

\item{ci}{Numerical vector of credible intervals.}
}
\value{
List of dataframes. Element \code{mean} returns the mean values for 
each variable at every date. Element \code{quantile} returns the 
corresponding quantiles based on input \code{ci}.
}
\description{
Summary statistics if a forecast object.
}
