% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{translate_i0prop_to_i0}
\alias{translate_i0prop_to_i0}
\title{Converts `i0.prop` into incidence matrix `i0`}
\usage{
translate_i0prop_to_i0(prms)
}
\arguments{
\item{prms}{List of model parameters}
}
\value{
List of model parameters
}
\description{
Converts `i0.prop` into incidence matrix `i0`
}
\keyword{internal}
