% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create an `amrem` model object.}
\usage{
create(prms, name = "no_name")
}
\arguments{
\item{prms}{List. Model parameters.}

\item{name}{String. Name of the model.}
}
\value{
List
}
\description{
Create an `amrem` model object.
}
\examples{
 1+1 #TODO

}
