% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_data_type}
\alias{fit_data_type}
\title{Type of data fitted.}
\usage{
fit_data_type(data)
}
\arguments{
\item{data}{Dataframe of observed data.}
}
\value{
Named logical vector. Element is TRUE if data type is fitted.
}
\description{
Type of data fitted.
}
\keyword{internal}
