% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast epidemic.}
\usage{
forecast(fitobj, simulate, prm.fcst, silent = TRUE)
}
\arguments{
\item{fitobj}{List. Object as returned by the function \code{fit()}.}

\item{simulate}{Logical. If \code{FALSE}, will simply use the posterior
simulations of the fit object. This implicitly constrains the time horizon
and the number of forecast trajectories.}

\item{prm.fcst}{List. Forecast parameters.}

\item{silent}{Logical. Display informational messages if \code{TRUE}. 
Default is \code{FALSE}.}
}
\value{
SOMETHING
}
\description{
Forecast epidemic.
}
