% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_data_names}
\alias{get_data_names}
\title{Helper function.}
\usage{
get_data_names(data.type, nag)
}
\arguments{
\item{data.type}{String.}

\item{nag}{Integet for the number of age groups.}
}
\description{
Helper function.
}
\keyword{internal}
