% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helper_generate_R}
\alias{helper_generate_R}
\title{Helper to generate a contact matrix}
\usage{
helper_generate_R(A, r0 = 1, decay = 0.1, min = 0.1, max = 0.4)
}
\arguments{
\item{A}{Integer. Number of age groups}

\item{r0}{}

\item{decay}{}

\item{min}{}

\item{max}{}
}
\value{
A contact matrix.
}
\description{
Helper to generate a contact matrix
}
\keyword{internal}
