% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{helper_plot_post}
\alias{helper_plot_post}
\title{Helper function to plot posteriors.}
\usage{
helper_plot_post(df, dfs, ncol = NULL, true.val = NULL)
}
\arguments{
\item{df}{dataframe of posterior data.}

\item{dfs}{dataframe of summary statistics.}

\item{ncol}{number of columns for faceted plots.}
}
\value{
a ggplot object.
}
\description{
Helper function to plot posteriors.
}
\keyword{internal}
