% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{helper_priors_vec}
\alias{helper_priors_vec}
\title{Helper function for prior vectors}
\usage{
helper_priors_vec(varname, nag, priors.dist, n.priors)
}
\arguments{
\item{varname}{String. Variable name.}

\item{nag}{Integer. Number of age groups.}

\item{priors.dist}{List defining prior distributions.}

\item{n.priors}{Integer. Number of prior samples.}
}
\description{
Helper function for prior vectors
}
\keyword{internal}
