% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit_post}
\alias{plot_fit_post}
\title{Plot fitted posteriors}
\usage{
plot_fit_post(fitobj, ci = 0.95, true.values = NULL)
}
\arguments{
\item{fitobj}{List. Object as returned by \code{fit()}.}

\item{ci}{(Optional) Numeric. Width (between 0 and 1) of the credible interval to plot. Default to 0.95.}
}
\value{
A list of ggplot objects. One element per fitted parameters.
}
\description{
Plot fitted posteriors
}
\examples{
# TO DO

}
