% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate an epidemic.}
\usage{
simulate(obj, check.prms = TRUE, ww = TRUE, hosp = TRUE, testpos = TRUE)
}
\arguments{
\item{obj}{List representing the `amrem` object as returned by \code{amrem::create()}.}

\item{check.prms}{Logical. Check input parameters? Default = TRUE.}

\item{ww}{A boolean to switch on/off fecal shedding into wastewater.}

\item{hosp}{A boolean to switch on/off hospital admissions.}

\item{testpos}{A boolean to switch on/off clinical test positivity.}
}
\value{
Dataframe of simulated variables.
}
\description{
Simulate an epidemic.
}
\examples{
prms = example_model_prms()
obj = amrem::create(prms = prms, name = 'foo')
s = simulate(obj)
head(s)

}
