% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate an epidemic.}
\usage{
simulate(prms, check.prms = TRUE)
}
\arguments{
\item{prms}{List of model parameters}

\item{check.prms}{Logical. Check input parameters? Default = TRUE.}
}
\value{
Dataframe of simulated variables.
}
\description{
Simulate an epidemic.
}
\examples{
1 + 2


}
